requirements_clustering.txt
numpy>=1.24.0
scikit-learn>=1.3.0
umap-learn>=0.5.4
matplotlib>=3.7.0
plotly>=5.17.0
